feat: picture Utils 분리
